Infra : 파이프 라인에 key 주입
