--- src/main/java/com/now_here5/now_here/infra/notification/service/FCMInitializer.java
@@ -1,33 +1,0 @@
-package com.now_here5.now_here.infra.notification.service;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import jakarta.annotation.PostConstruct;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.stereotype.Service;
-
-import java.io.IOException;
-
-@Service
-@Slf4j
-public class FCMInitializer {
-
-    private static final String FIREBASE_CONFIG_PATH = "nowhere-fe86a-firebase-adminsdk-j3py1-70a20e7e35.json";
-
-    @PostConstruct
-    public void initialize() {
-        try {
-            GoogleCredentials googleCredentials = GoogleCredentials
-                    .fromStream(new ClassPathResource(FIREBASE_CONFIG_PATH).getInputStream());
-            FirebaseOptions options = new FirebaseOptions.Builder()
-                    .setCredentials(googleCredentials)
-                    .build();
-            FirebaseApp.initializeApp(options);
-        } catch (IOException e) {
-            log.info(">>>>>>>>FCM error");
-            log.error(">>>>>>FCM error message : " + e.getMessage());
-        }
-    }
-}
