+++ src/main/resources/application.properties
@@ -39,3 +39,6 @@ spring.mail.host=${MAIL_HOST}
 spring.mail.port=${MAIL_PORT}
 spring.mail.username=${MAIL_USERNAME}
 spring.mail.password=${MAIL_PASSWORD}
+
+# FCM properties
+fcm.credentials.path=${FCM_CREDENTIALS_PATH}
